Добавил сборку тестов (пока есть конфликт с главным main ядра)
